refactor(#3479): shrink diagnostics.rs below giant threshold
Verbatim-extract the diagnostic report builders and the snapshot/queue
normalization + runtime-labeling helpers they consume from the registered
giant `commands/diagnostics.rs` (1022 prod LoC) into a cohesive child module,
graduating the file from the giant-file registry/baseline.

Structure: diagnostics.rs -> diagnostics/ directory module.
- commands/diagnostics/mod.rs (389 prod LoC): metrics/health/sessions/status/
  inflight/queue/adk-phase/debug slash commands + the Gemini session helpers
  (run_blocking_gemini_command, format_gemini_session_report). Re-exports the
  report builders so existing callers keep resolving `diagnostics::build_*`
  (commands/skill.rs) and the commands/mod.rs re-export (commands/text_commands.rs).
- commands/diagnostics/reports.rs (651 prod LoC): build_health_report,
  build_status_report, build_inflight_report, build_queue_report(_sync) plus
  shorten_session_identifier, inflight_runtime_label, fetch_raw_provider_session_id,
  and the pending-queue normalization helpers. The runtime_label_tests unit test
  moved with it.

Behavior-preserving: the moved bodies are byte-identical to the originals; only
the module path depth (super::super:: -> super::super::super:: for the three
discord-level callsites) and the report-builder re-export visibility changed.

Graduation metadata:
- scripts/audit_maintainability_giant_baseline.toml: drop the diagnostics.rs
  ratchet entry (decomposed below threshold; health.rs precedent).
- scripts/giant_file_registry.toml: remove the ghost diagnostics.rs from
  `grandfathered` and `grandfathered_baseline_paths` (config.rs/health.rs
  precedent).
- docs/agent-maintenance/change-surfaces.md: drop diagnostics.rs from the
  giant brace group + add a #3479 refresh note for the new submodule.
- ARCHITECTURE.md + generated inventory regenerated.

Gates: cargo check (0 errors), cargo test --lib diagnostics (5 passed),
audit_maintainability.py --check (rc 0), generate_inventory_docs.py (rc 0),
ci-script-checks.sh (rc 0; discord-outbound-migration.md soft warning only).
