+++ ARCHITECTURE.md
@@ -358,6 +358,9 @@ src/
 │   │   └── session.rs
 │   ├── discord/
 │   │   ├── commands/
+│   │   │   ├── diagnostics/
+│   │   │   │   ├── mod.rs
+│   │   │   │   └── reports.rs
 │   │   │   ├── inspect/
 │   │   │   │   ├── formatting.rs
 │   │   │   │   ├── mod.rs
@@ -372,7 +375,6 @@ src/
 │   │   │   ├── command_policy.rs
 │   │   │   ├── config.rs
 │   │   │   ├── control.rs
-│   │   │   ├── diagnostics.rs
 │   │   │   ├── fast_mode.rs
 │   │   │   ├── goals.rs
 │   │   │   ├── help.rs
